lt-test-backend/requirements.txt
# Django
django==3.1.1
djangorestframework==3.11.1
django-filter==2.3.0

# Other
Pillow==7.2.0
psycopg2-binary==2.8.6
markdown==3.2.2